backend/ai/voice_emotion_model.py
"""
Voice Emotion Recognition Model

This module provides functionality to detect emotions from voice recordings
using a pre-trained deep learning model.
"""

import numpy as np
import librosa

# Mock implementation of a voice emotion detection model
class VoiceEmotionDetector:
    def __init__(self, model_path=None):
        """
        Initialize the voice emotion detector.
        
        Args:
            model_path: Path to the pre-trained model weights
        """
        self.emotions = ['angry', 'calm', 'disgust', 'fear', 'happy', 'neutral', 'sad', 'surprise']
        self.model_loaded = True
        print("Voice emotion detection model loaded successfully")
        
    def extract_features(self, audio_path, sr=22050):
        """
        Extract audio features from a voice recording.
        
        Args:
            audio_path: Path to the audio file
            sr: Sample rate
            
        Returns:
            Dictionary of extracted features
        """
        try:
            # Load audio file
            y, sr = librosa.load(audio_path, sr=sr)
            
            # Extract features
            mfccs = librosa.feature.mfcc(y=y, sr=sr, n_mfcc=13)
            chroma = librosa.feature.chroma_stft(y=y, sr=sr)
            mel = librosa.feature.melspectrogram(y=y, sr=sr)
            contrast = librosa.feature.spectral_contrast(y=y, sr=sr)
            tonnetz = librosa.feature.tonnetz(y=librosa.effects.harmonic(y), sr=sr)
            
            # Compute statistics
            features = {
                'mfcc_mean': np.mean(mfccs, axis=1),
                'mfcc_std': np.std(mfccs, axis=1),
                'chroma_mean': np.mean(chroma, axis=1),
                'mel_mean': np.mean(mel, axis=1),
                'contrast_mean': np.mean(contrast, axis=1),
                'tonnetz_mean': np.mean(tonnetz, axis=1)
            }
            
            return features
            
        except Exception as e:
            print(f"Error extracting features: {str(e)}")
            return None
        
    def detect_emotion(self, audio_path):
        """
        Detect the emotion from a voice recording.
        
        Args:
            audio_path: Path to the audio file
            
        Returns:
            Dictionary with detected emotion and confidence score
        """
        # This would normally extract features and use the model to predict
        # For now, returning a mock result
        
        # Mock feature extraction
        features = self.extract_features(audio_path) if audio_path else None
        
        # Mock prediction
        # In a real implementation, this would pass the features to a neural network
        emotion_idx = np.random.randint(0, len(self.emotions))
        confidence = np.random.uniform(0.6, 0.95)
        
        return {
            "emotion": self.emotions[emotion_idx],
            "confidence": float(confidence),
            "features_extracted": features is not None
        }
        
    def detect_emotions_from_conversation(self, audio_segments):
        """
        Detect emotions from a conversation by analyzing multiple audio segments.
        
        Args:
            audio_segments: List of paths to audio segments
            
        Returns:
            List of emotions detected for each segment
        """
        results = []
        
        for segment in audio_segments:
            emotion_result = self.detect_emotion(segment)
            results.append(emotion_result)
            
        return results

# Example usage
if __name__ == "__main__":
    detector = VoiceEmotionDetector()
    # This would normally use a real audio file
    result = detector.detect_emotion(None)
    print(f"Detected emotion: {result['emotion']} with confidence {result['confidence']:.2f}")
